+++ apps/agora/app/e2e/helpers/comparison-tool.ts
@@ -3,18 +3,16 @@ import {
   ComparisonToolConfig,
   ComparisonToolConfigFilter,
 } from '@sagebionetworks/agora/api-client';
+import { DEFAULT_PAGE_SIZE } from '@sagebionetworks/explorers/constants';
 import { expectComparisonToolTableLoaded } from '@sagebionetworks/explorers/testing/e2e';
+import { camelCase } from 'lodash';
 import { baseURL } from '../../playwright.config';
 import {
   COMPARISON_TOOL_API_PATHS,
   COMPARISON_TOOL_CONFIG_PATH,
-  COMPARISON_TOOL_DEFAULT_PAGE_SIZE,
   COMPARISON_TOOL_DEFAULT_SORTS,
   COMPARISON_TOOL_PATHS,
 } from './constants';
-
-const snakeToCamel = (value: string): string =>
-  value.replace(/_([a-z])/g, (_match, char: string) => char.toUpperCase());
 
 export const navigateToComparison = async (
   page: Page,
@@ -43,6 +41,7 @@ export const fetchComparisonToolData = async <T>(
   page: Page,
   name: string,
   categories: string[] = [],
+  extraParams?: URLSearchParams,
 ): Promise<T> => {
   const params = new URLSearchParams();
   params.append('itemFilterType', 'exclude');
@@ -55,6 +54,12 @@ export const fetchComparisonToolData = async <T>(
   for (const sort of defaultSort) {
     params.append('sortFields', sort.field);
     params.append('sortOrders', sort.order.toString());
+  }
+
+  if (extraParams) {
+    for (const [key, value] of extraParams) {
+      params.append(key, value);
+    }
   }
 
   const response = await page.request.get(`${baseURL}/api/v1/${COMPARISON_TOOL_API_PATHS[name]}`, {
@@ -83,26 +88,19 @@ const fetchFilteredPageCount = async (
   filter: ComparisonToolConfigFilter,
   value: string,
 ): Promise<number> => {
-  const params = new URLSearchParams();
-  params.append('itemFilterType', 'exclude');
-  params.append('pageSize', COMPARISON_TOOL_DEFAULT_PAGE_SIZE.toString());
-
-  // sortFields and sortOrders are required by the API
-  const defaultSort = COMPARISON_TOOL_DEFAULT_SORTS[name];
-  for (const sort of defaultSort) {
-    params.append('sortFields', sort.field);
-    params.append('sortOrders', sort.order.toString());
-  }
+  const extraParams = new URLSearchParams();
+  extraParams.append('pageSize', DEFAULT_PAGE_SIZE.toString());
 
   // The config's query_param_key targets the frontend route; the API filters on the
   // camelCased data_key (e.g. maximum_clinical_trial_phase -> maximumClinicalTrialPhase).
-  params.append(snakeToCamel(filter.data_key), value);
+  extraParams.append(camelCase(filter.data_key), value);
 
-  const response = await page.request.get(`${baseURL}/api/v1/${COMPARISON_TOOL_API_PATHS[name]}`, {
-    params,
-  });
-  expect(response.ok()).toBeTruthy();
-  const data = (await response.json()) as { page: { totalPages: number } };
+  const data = await fetchComparisonToolData<{ page: { totalPages: number } }>(
+    page,
+    name,
+    [],
+    extraParams,
+  );
   return data.page.totalPages;
 };
