+++ libs/explorers/services/src/lib/app-storage.constants.ts
@@ -1,4 +1,4 @@
+export { DEFAULT_PAGE_SIZE, VALID_PAGE_SIZES } from '@sagebionetworks/explorers/constants';
+
 export const HIDE_VISUALIZATION_OVERVIEW_KEY = 'hide_visualization_overview';
 export const PAGE_SIZE_KEY = 'comparison_tool_page_size';
-export const VALID_PAGE_SIZES: readonly number[] = [10, 25, 50];
-export const DEFAULT_PAGE_SIZE = VALID_PAGE_SIZES[0];
